added DeletePost tests for controller -- all tests pass -- todo: test only 1 post gets deleted

+++ backend/src/main/java/be/vdab/restcontrollers/PostController.java
@@ -46,6 +46,9 @@ public class PostController {
 
     @PostMapping("delete")
     public ResponseEntity<String> deletePost(@RequestBody Post post){
+        if(post == null){
+            return new ResponseEntity<>("Delete failed", new HttpHeaders(), HttpStatus.CONFLICT);
+        }
         boolean result = postService.deletePostByID(post.getId());
         if(result){
             return new ResponseEntity<>("Delete success", new HttpHeaders(), HttpStatus.OK);
